Migrate Data tests to munit
